Add missing yield() support
The yield function was added to the Arduino core to allow for
cooperative scheduling, but in the case of the AVR core, is simply
defined as a void function to prevent error in code that expects it,
such as the FastLED library. This patch removes the "error: 'yield' was
not declared in this scope" compile failure from the digispark and
digispark pro cores.